Add crude JSON:API parsing code.
Connects to our local server that returns Books in a JSON:API format. Decodes the JSON structure, then processes it into a local in-memory object graph.

This is not production code. Rather, I'm simply trying to stand up an apples-to-apples comparison of how JSON:API vs SQLite from the server affects client side processing and performance.